+++ Network/API/Telegram/Bot/Object/Sender.hs
@@ -15,8 +15,8 @@ import "lens" Control.Lens (Lens')
 import "tagged" Data.Tagged (Tagged)
 import "text" Data.Text (Text)
 
-import Network.API.Telegram.Bot.Property.Persistable
-	(Persistable (Payload, payload, endpoint), Capacity (Fetch))
+import Network.API.Telegram.Bot.Property.Identifiable (Identifiable (Identificator, identificator))
+import Network.API.Telegram.Bot.Property.Persistable (Persistable (Payload, payload, endpoint), Capacity (Fetch))
 
 data Sender
 	= Bot Int (Maybe Text) Text (Maybe Text) (Maybe Text)
@@ -35,11 +35,14 @@ instance FromJSON Sender where
 		>>= bool (sender v User) (sender v Bot) where
 
 		sender :: Object -> Whom -> Parser Sender
-		sender v f = f <$> v .: "id"
-			<*> v .:? "username"
-			<*> v .: "first_name"
-			<*> v .:? "last_name"
+		sender v f = f <$> v .: "id" <*> v .:? "username"
+			<*> v .: "first_name" <*> v .:? "last_name"
 			<*> v .:? "language_code"
+
+instance Identifiable Sender where
+	type instance Identificator Sender = Int
+	identificator (Bot i _ _ _ _) = i
+	identificator (User i _ _ _ _) = i
 
 instance Persistable 'Fetch Sender where
 	type instance Payload 'Fetch Sender = Tagged ('Fetch Sender) ()