+++ Network/API/Telegram/Bot/Object/Update/Message/Origin.hs
@@ -11,6 +11,7 @@ import "base" Text.Show (Show)
 import "text" Data.Text (Text)
 
 import Network.API.Telegram.Bot.Object.Sender (Sender)
+import Network.API.Telegram.Bot.Property.Identifiable (Identifiable (Identificator, identificator))
 
 data Origin
 	= Private Int64 Sender
@@ -29,3 +30,10 @@ instance FromJSON Origin where
 			("supergroup" :: Text) -> Supergroup <$> c .: "id" <*> c .: "title" <*> v .: "from"
 			("channel" :: Text) -> Channel <$> c .: "id" <*> c .: "title"
 			_ -> fail "Type of chat is not defined"
+
+instance Identifiable Origin where
+	type instance Identificator Origin = Int64
+	identificator (Private i _) = i
+	identificator (Group i _ _) = i
+	identificator (Supergroup i _ _) = i
+	identificator (Channel i _) = i