+++ CHANGELOG.md
@@ -92,3 +92,4 @@
 * Define `Persistable` instances for `Poll`, `Audio`, `Document`, `Video`, `Voice` datatypes
 * Define `Group` datatype for representation basic groups and supergroups
 * Remove `Object` type family to distribute `Accessible` and `Identifiable` instances
+* Put `Identificator` type family into `Identifiable` type class